Implement memory efficient "find references" with bloom filters.
This commit implements `textDocument/references` with the following
hard constraints:

- requests should respond within 100-200ms like for definition requests
- memory usage for any pre-computed index should not exceed 10-100mb
- computing an index should be fast, no more than 5-10 seconds for large workspaces
- the index should be easy to incrementally update as files get recompiled

To achieve these goals, we build an index that is a map keyed by
SemanticDB file paths with bloom filters as values.

```scala
val index = Map[Path, BloomFilter[Symbol]]
```

We compute the index by walking all SemanticDB files and entering all
referenced symbol into the bloom filter. To respond to "find references"
requests, we iterate through the index, filter out paths that are
guaranteed not to reference the requested symbol, and read the
SemanticDB files from disk which might reference the requested symbol.

I evaluated this approach in the Prisma repo and it seems we comfortably
beat the hard constraints:
- <1 second to index SemanticDB files for 80k loc on a cold server
- 700kb memory to store index of all SemanticDB symbol references
- most requests are responded in 3-100ms, and very frequently referenced
  symbols like `String` still respond in ~700ms.

By supporting "find references", we need to change our compilation
strategy to regularly compile all sources in the workspace that depend
on the open file. Previously, if we focused on a file A, we only
compiled the build target that contained A. Now, we also compile build
targets that depend on A. It's possible to use the old compilation
strategy with `-Dmetals.inverse-compile=false`, but if that option
becomes popular then we should discuss how to improve the default
mode.