Merge pc module into mtags and rename pc-interfaces into mtags-interfaces.
In my experience, sbt does not scale well with a growing number of
projects. If we were using pants or bazel I would prefer to write small
composable targets, but in sbt I think it's better to write large
projects with a lot of bundled functionality.

I want to avoid repeating the same mistake we made in the Scalameta
build, which has ~30 micro-projects (which becomes >100 modules when you
start cross-publishing for 2.11/2.12 and Scala.js/Native). In theory,
it's good to have small and self-contained components so users can
depend only on the functionality they need. However, my experience with
micro-projects in sbt is that they greatly complicate the build:

- publishing artifacts becomes slow, causing more frequent timeouts
  during releases.
- the sbt build becomes slower to work in and uses more memory, making
  the build less accessible for contributors with underpowered computers.
- small helper methods ends up getting duplicated across micro-libraries
  because the cost of creating a new shared project for those helper
  methods is too high in sbt.